backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.3.0",
    "drf-simplejwt>=4.4.1",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "python-decouple>=3.8",
]
